+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Folding@home Web Control Changelog
 =================================
 
+## v8.3.5
+ - Fix Stats display with long user name. #138
+ - Show project number with RCG in WU details view. #131
+
 ## v8.3.4
  - Beta release